Apply nodeAffinity when mount=none
* Apply nodeAffinity when mount=none

When mount is `all` we target the nfs workers explicitly with a
`nodeSelector`, which forces the scheduler to only use the nfs workers.

Currently when mount is `none` we do nothing, which allows the scheduler
to place the jobs on any worker.

This works fine but has 2 downsides:
1. Jobs that don't use mounts use capacity which may be required by jobs
   that do use mounts, making them un-schedulable.

2. Jobs that don't use mounts are affected by nfs issues (e.g. when
   workers lock up)

This change introduces a patch which applies a nodeAffinity,
influencing the scheduler to place the workloads on a non nfs worker
node.

The policy is configured as `preferredDuringSchedulingIgnoredDuringExecution`,
rather than `requiredDuringSchedulingIgnoredDuringExecution`,
this enables a casual migration of jobs.

Over time the non-nfs workers should fill up and then the overflow will
be scheduled onto the nfs workers, during this time the cluster can be
re-scaled towards 'stateless' worker nodes as desired.

The primary effect this should have is to improve the experience for
maintainers regarding the stability of their jobs.